New controller for server databases page
